marquee has sub-strategies: selection and operation
